Version 1.5
NF: The yankring now respects the cpoptions setting, if 'y' is included and you press '.', the previous yank command is executed and added to the yankring.  You can also add this behaviour by setting this in your |vimrc|:
       let g:yankring_dot_repeat_yank = 1

NF: Duplicates will not be added to the yankring by default.  If a duplicate is found, the element will be moved to the top of the yankring.  This can be controlled by setting this in your |vimrc|:
       let g:yankring_ignore_duplicate = 0 (1 is default)

BF: Regression from version 1.4, the '.' operator may incorrectly insert garbage.