Version 7.0
NF: Added support for the / and ? motions so that y/search is supported (Vasilii Pascal).
NF: When the YankRing window is displayed (or updated) an additional check is made against the default register.  If it has changed since the YankRing recorded it, the value will be added to the history.
NF: Added support for more motions h, j, k, l, H, M, L, ^, 0, -, +, _.  And a pile of g motions g_, g^, gm, g$, gk, gj, gg, ge, gE.
NF: The YankRing window will display a message it is operating in a limited mode if not using Vim 7.2 or the correct patch level.
BF: Correction to some internal code which could lead to an endless loop (John Beckett).
BF: Opening and closing the YankRing window with "set report=0" reported "1 line less" messages (Bill McCarthy).
BF: Changed the default value of g:yankring_paste_check_default_buffer to check if the default paste buffer has changed when pressing 'p'.  For example, if a plugin has changed the default registers it will be pasted rather than the top item from the YankRing.
BF: YRMapsDelete did not remove all the maps created by the YankRing.
BF: Under particular circumstances, yanking text with embedded @ characters were not properly stored and retrieved from the YankRing (Andrew Long).
BF: Changed to use xmaps instead of vmaps so that the maps only work in visual mode and not select mode (David Liang).