Version 8.0
NF: Changed the implementation of YRGetSearch() (David Liang).
BF: Under some unknown circumstances, the yankring can fail to change to the correct buffer.  Put in code to double check and abort.
BF: Yanking and pasting a line which ends in a backslash resulted in the backslash being replaced by "@@@".
BF: When repeating a command (".") which used any of the zap keys (f,F,t,T,/,?) you were prompted again for the string to match on (Vasilii Pascal).