Version 14.0
NF: The YankRing window supports 1-9 to choose to paste those items for quicker access to the top 9 items. 

NF: The YankRing now maps the @ key to run macros.  Not all actions performed during the macro are recorded by the YankRing as a number of items had to be unmapped to support the replay of actions which can prompt the user for input (i.e. t and f) (Asis Hallab). 

BF: When flipping between applications the system clipboard was not added to the YankRing in all cases. 

BF: The YankRing could report E121: Undefined variable: g:yankring_manual_clipboard_check (Thilo Six). 

BF: The YankRing could report: E132: Function call depth is higher than 'maxfuncdepth'.  When executing a YRClear or YRPop or when flipping between applications (Marcin Szamotulski). 

BF: Strange behaviour when opening the YankRing with :e [YankRing] instead of :YRShow (Marcin Szamotulski).